last of changes from Wednesday meeting

+++ kern/include/proc.h
@@ -44,7 +44,7 @@ struct vnode;
 struct _file{
    struct vnode *vn;  /*this is the vnode pointing to that perticular file*/
    int fd;            /*this is the index of that vnode within the file t able*/
-   int seek;          /*this is keeping track of the "last modified" point in the file*/
+   off_t seek;          /*this is keeping track of the "last modified" point in the file*/
    struct spinlock spin;
    const char *file_name;
 };
@@ -95,7 +95,8 @@ void set_file_vnode(int fd , struct vnode *vn);
 void set_seek (int fd , int seek);
 
 /**/
-int get_seek (int fd);
+off_t get_seek (int fd);
+
 
 void set_file_name(int file_descriptor, const char *file_name);
